+++ src/__test__/test_production_2.py
@@ -4,112 +4,8 @@ from copy import deepcopy
 
 from graph import Graph
 from model import Node, NodeAttrs, EdgeAttrs, Edge
-from production import P1, P2
 
-
-class TestProduction1(unittest.TestCase):
-    def test_production_can_be_applied_on_lhs_graph(self):
-        graph = Graph()
-
-        node_1 = Node(NodeAttrs('v', 0, 0, False))
-        node_2 = Node(NodeAttrs('v', 1, 0, False))
-        node_3 = Node(NodeAttrs('v', 1, 1, False))
-        node_4 = Node(NodeAttrs('v', 0, 1, False))
-        nodes = [node_1, node_2, node_3, node_4]
-
-        graph.add_node_collection(nodes)
-
-        graph.add_q_hyperedge(nodes, EdgeAttrs('q', True))
-
-        for node_a, node_b in it.pairwise(nodes + [node_1]):
-            graph.add_edge(Edge(node_a.handle, node_b.handle, EdgeAttrs(kind='e', flag=False)))
-
-        p1 = P1()
-        mapping_gen = graph.generate_subgraphs_isomorphic_with(p1.get_lhs())
-        self.assertTrue(any(p1.is_mapping_feasible(graph, mapping) for mapping in mapping_gen))
-
-    def test_production_can_be_applied_on_more_complex_graph(self):
-        graph = Graph()
-
-        node_1 = Node(NodeAttrs('v', 0, 0, False))
-        node_2 = Node(NodeAttrs('v', 0.5, 0, False))
-        node_3 = Node(NodeAttrs('v', 1, 0, False))
-        node_4 = Node(NodeAttrs('v', 1, 1, False))
-        node_5 = Node(NodeAttrs('v', 0.5, 1, False))
-        node_6 = Node(NodeAttrs('v', 0, 1, False))
-        nodes = [node_1, node_2, node_3, node_4, node_5, node_6]
-
-        graph.add_node_collection(nodes)
-        graph.add_q_hyperedge((node_1, node_2, node_5, node_6), EdgeAttrs('q', True))
-        graph.add_q_hyperedge((node_2, node_4, node_3, node_5), EdgeAttrs('q', True))
-
-        graph.add_edge(Edge(node_2.handle, node_5.handle, EdgeAttrs(kind='e', flag=False)))
-        for node_1, node_2 in it.pairwise(nodes + [node_1]):
-            graph.add_edge(Edge(node_1.handle, node_2.handle, EdgeAttrs(kind='e', flag=False)))
-
-        p1 = P1()
-        mapping_gen = graph.generate_subgraphs_isomorphic_with(p1.get_lhs())
-        self.assertTrue(any(p1.is_mapping_feasible(graph, mapping) for mapping in mapping_gen))
-
-    def test_production_cannot_be_applied_if_R_false(self):
-        graph = Graph()
-
-        node_1 = Node(NodeAttrs('v', 0, 0, False))
-        node_2 = Node(NodeAttrs('v', 1, 0, False))
-        node_3 = Node(NodeAttrs('v', 1, 1, False))
-        node_4 = Node(NodeAttrs('v', 0, 1, False))
-        nodes = [node_1, node_2, node_3, node_4]
-
-        graph.add_node_collection(nodes)
-
-        graph.add_q_hyperedge((node_1, node_2, node_3, node_4), EdgeAttrs('q', False))
-
-        for node_a, node_b in it.pairwise(nodes + [node_1]):
-            graph.add_edge(Edge(node_a.handle, node_b.handle, EdgeAttrs(kind='e', flag=False)))
-
-        p1 = P1()
-        mapping_gen = graph.generate_subgraphs_isomorphic_with(p1.get_lhs())
-        self.assertFalse(any(p1.is_mapping_feasible(graph, mapping) for mapping in mapping_gen))
-
-    def test_production_cannot_be_applied_if_edge_types_not_match(self):
-        graph = Graph()
-
-        node_1 = Node(NodeAttrs('v', 0, 0, False))
-        node_2 = Node(NodeAttrs('v', 1, 0, False))
-        node_3 = Node(NodeAttrs('v', 1, 1, False))
-        node_4 = Node(NodeAttrs('v', 0, 1, False))
-        nodes = [node_1, node_2, node_3, node_4]
-
-        graph.add_node_collection(nodes)
-        graph.add_q_hyperedge((node_1, node_2, node_3, node_4), EdgeAttrs('q', True))
-
-        graph.add_edge(Edge(node_4.handle, node_1.handle, EdgeAttrs(kind='q', flag=True)))
-        for node_a, node_b in it.pairwise(nodes):
-            graph.add_edge(Edge(node_a.handle, node_b.handle, EdgeAttrs(kind='e', flag=False)))
-
-        p1 = P1()
-        mapping_gen = graph.generate_subgraphs_isomorphic_with(p1.get_lhs())
-        self.assertFalse(any(p1.is_mapping_feasible(graph, mapping) for mapping in mapping_gen))
-
-    def test_production_cannot_be_applied_if_exists_hanging_node(self):
-        graph = Graph()
-
-        node_1 = Node(NodeAttrs('v', 0, 0, False))
-        node_2 = Node(NodeAttrs('v', 1, 0, False))
-        node_3 = Node(NodeAttrs('v', 1, 1, False))
-        node_4 = Node(NodeAttrs('v', 0, 1, True))
-        nodes = [node_1, node_2, node_3, node_4]
-
-        graph.add_node_collection(nodes)
-        graph.add_q_hyperedge((node_1, node_2, node_3, node_4), EdgeAttrs('q', True))
-
-        for node_a, node_b in it.pairwise(nodes + [nodes[0]]):
-            graph.add_edge(Edge(node_a.handle, node_b.handle, EdgeAttrs(kind='e', flag=False)))
-
-        p1 = P1()
-        mapping_gen = graph.generate_subgraphs_isomorphic_with(p1.get_lhs())
-        self.assertFalse(any(p1.is_mapping_feasible(graph, mapping) for mapping in mapping_gen))
-
+from production import P2
 
 class TestProduction2(unittest.TestCase):
     def test_production_can_be_applied_on_lhs_graph(self):
@@ -257,7 +153,6 @@ class TestProduction2(unittest.TestCase):
         p2 = P2()
         mapping_gen = graph.generate_subgraphs_isomorphic_with(p2.get_lhs())
         self.assertFalse(any(p2.is_mapping_feasible(graph, mapping) for mapping in mapping_gen))
-
 
 if __name__ == '__main__':
     unittest.main()
